digital_twin_logic/backend/requirements.txt
# Backend Python dependencies for FastAPI + MQTT + Supabase
fastapi>=0.95
uvicorn[standard]>=0.22
python-dotenv>=1.0
supabase>=2.22.0
httpx>=0.28.1
paho-mqtt>=1.6
psycopg2-binary>=2.9.11
pydantic>=2.12.3

# API documentation
annotated-types>=0.7.0

# Database and authentication
cryptography>=46.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=7.0
pytest-asyncio>=0.22
httpx>=0.24